Fix display when backspace exponential number
Display would reset to 0 when backspacing a scientifically
notated number.  Also, a number would return an error when
clicking 'equals' when not performing a calculation.  This is
because 'number1' is undefined when not performing a calculation.
Also, 'operator' was undefined .  Made the findResult equation
return number2 when number1 is undefined